pyproject.toml
[project]
name = "max-stats"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=1.35.1",
    "typer-slim>=0.20.0",
]
